simhash
This is a fast testing for simhash haiming distance.
The test is pass in the first stage as this algorithm can realize the similarity exist between different paper.
Working code is independent from master branch, been written in simhash.class, the development will also show not interruption on other function
this class using pdfbox as import JAR to handle the pdf to text, but the performance of pdfbox is not so well. Better JAR library is required to increase its work efficiency.